+++ index.html
@@ -195,8 +195,7 @@
       </ul>
     </div>
   </footer>
-  <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 
   <!-- Шаблон изображения случайного пользователя -->
   <template id="picture">
