+++ js/random-arrow-element.js
@@ -1,0 +1,4 @@
+import {getRandomInteger} from './random-integer.js';
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomArrayElement};
