+++ js/random-number.js
@@ -1,0 +1,5 @@
+function getRandomNumber(maxValue) {
+  return Math.floor(Math.random() * (maxValue + 1));
+}
+
+export {getRandomNumber};
